(UI Optimization and removing Extra Code) All remaining Screens Are optimized.
