+++ lib/ui/screens/pages/account/address_book/add_address/add_address_screen.dart
@@ -1,6 +1,4 @@
 import 'package:danapaniexpress/core/common_imports.dart';
-import 'package:danapaniexpress/ui/screens/pages/account/address_book/add_address/add_address_mobile.dart';
-import 'package:danapaniexpress/ui/screens/pages/account/address_book/address_book_mobile.dart';
 
 class AddAddressScreen extends StatelessWidget {
   const AddAddressScreen({super.key});
