+++ lib/ui/screens/pages/categories/other_products/other_products_screen_mobile.dart
@@ -7,40 +7,12 @@ class OtherProductsScreenMobile extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final otherProductsController = Get.put(OtherProductsController());
-    final navigation = Get.find<NavigationController>();
-    final home = Get.find<HomeController>();
     final products = Get.find<ProductsController>();
 
     return Obx(() {
       final screenType = otherProductsController.productScreenType.value;
-
-      final coverImage = screenType == ProductsScreenType.FEATURED
-          ? home.coverImages.value!.featured
-          : screenType == ProductsScreenType.FLASHSALE
-          ? home.coverImages.value!.flashSale
-          : screenType == ProductsScreenType.POPULAR
-          ? home.coverImages.value!.popular
-          : null;
-
-      final screenName = screenType == ProductsScreenType.FEATURED
-          ? AppLanguage.featuredProductStr(appLanguage)
-          : screenType == ProductsScreenType.FLASHSALE
-          ? AppLanguage.flashSaleStr(appLanguage)
-          : screenType == ProductsScreenType.POPULAR
-          ? AppLanguage.popularProductsStr(appLanguage)
-          : null;
-      final productsList = screenType == ProductsScreenType.FEATURED
-          ? products.featuredProducts
-          : screenType == ProductsScreenType.FLASHSALE
-          ? products.flashSaleProducts
-          : products.popularProducts;
-      final loadStatus = screenType == ProductsScreenType.FEATURED
-          ? products.featuredStatus.value
-          : screenType == ProductsScreenType.FLASHSALE
-          ? products.flashSaleStatus.value
-          : products.popularStatus.value;
-
       final isLoadingMore = products.isLoadingMore.value;
+
       final hasMore = screenType == ProductsScreenType.FEATURED
           ? products.hasMoreFeaturedProducts.value
           : screenType == ProductsScreenType.FLASHSALE
@@ -49,50 +21,18 @@ class OtherProductsScreenMobile extends StatelessWidget {
 
       final reachedEnd = otherProductsController.reachedEndOfScroll.value;
 
-      return _buildUI(
-        otherProductsController,
-        navigation,
-        home,
-        coverImage,
-        screenName,
-        loadStatus,
-        productsList,
-        isLoadingMore,
-        hasMore,
-        reachedEnd,
-      );
+      return _buildUI(isLoadingMore, hasMore, reachedEnd, screenType);
     });
   }
 
-  Widget _buildUI(
-    otherProductsController,
-    navigation,
-    home,
-    coverImage,
-    screenName,
-    loadStatus,
-    productsList,
-    isLoadingMore,
-    hasMore,
-    reachedEnd,
-  ) {
+  Widget _buildUI(isLoadingMore, hasMore, reachedEnd, screenType) {
     return Column(
       children: [
         // Top Banner
-        _topBannerSection(
-          otherProductsController,
-          screenName,
-          coverImage,
-          navigation,
-        ),
+        OtherProductsTopBannerSection(screenType: screenType),
 
         // Product List
-        _productsList(
-          loadStatus,
-          productsList,
-          otherProductsController,
-          navigation,
-        ),
+        OtherProductList(screenType: screenType),
 
         // ✅ Bottom Message Section
         BottomMessagesSection(
@@ -103,73 +43,5 @@ class OtherProductsScreenMobile extends StatelessWidget {
         ),
       ],
     );
-  }
-
-  Widget _topBannerSection(
-    otherProductsController,
-    screenName,
-    coverImage,
-    navigation,
-  ) {
-    return AnimatedSwitcher(
-      duration: const Duration(milliseconds: 250),
-      child: otherProductsController.showTopHeader.value
-          ? TopImageHeader(title: screenName!, coverImage: coverImage ?? "")
-          : appBarCommon(
-              title: screenName!,
-              isBackNavigation: true,
-              isTrailing: true,
-              trailingIcon: icSearch,
-              trailingIconType: IconType.SVG,
-              trailingOnTap: () => navigation.gotoSearchScreen(),
-            ),
-    );
-  }
-
-  Widget _productsList(
-    loadStatus,
-    productsList,
-    otherProductsController,
-    navigation,
-  ) {
-    return loadStatus == ProductsStatus.LOADING
-        ? Expanded(child: loadingIndicator())
-        : loadStatus == ProductsStatus.FAILURE
-        ? Expanded(child: ErrorScreen())
-        : productsList.isEmpty
-        ? Expanded(
-            child: EmptyScreen(
-              icon: AppAnims.animEmptyBoxSkin(isDark),
-              text: AppLanguage.noProductsStr(appLanguage).toString(),
-            ),
-          )
-        : Expanded(
-            child: SingleChildScrollView(
-              controller: otherProductsController.scrollController,
-              physics: BouncingScrollPhysics(),
-              child: Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: MAIN_HORIZONTAL_PADDING,
-                  vertical: MAIN_VERTICAL_PADDING,
-                ),
-                child: ConstrainedBox(
-                  constraints: BoxConstraints(minHeight: size.height),
-                  child: StaggeredGrid.count(
-                    crossAxisCount: 2,
-                    mainAxisSpacing: MAIN_HORIZONTAL_PADDING,
-                    crossAxisSpacing: MAIN_HORIZONTAL_PADDING,
-                    children: List.generate(productsList.length, (index) {
-                      final product = productsList[index];
-                      return GestureDetector(
-                        onTap: () =>
-                            navigation.gotoProductDetailScreen(data: product),
-                        child: ProductItem(data: product),
-                      );
-                    }),
-                  ),
-                ),
-              ),
-            ),
-          );
   }
 }
