+++ lib/ui/screens/startup_screens/startup_main_screen/startup_main_screen.dart
@@ -32,16 +32,7 @@ class StartupMainScreen extends StatelessWidget {
 }
 
 Widget buildMobileUI() {
-  return StartupScreenLayoutMobile(
-    // image: EnvImages.imgWelcomeStartup,
-    // headingText: AppLanguage.startupWelcomeHeadingStr(appLanguage).toString(),
-    // subText: AppLanguage.startupWelcomeSubtextStr(appLanguage).toString(),
-    // buttonText: AppLanguage.getStartedStr(appLanguage).toString(),
-    // onTap: (){
-    //   print('onTap working');
-    //   JumpTo.gotoFastDeliveryStartupScreen();
-    // },
-  );
+  return StartupScreenLayoutMobile();
 }
 
 Widget buildTabletUI() {
